Organize endpoints via routers
